+++ agape/src/widgets/button.rs
@@ -83,14 +83,6 @@ impl Widget for Button {
         }
     }
 
-    fn children(&self) -> Vec<&dyn Widget> {
-        vec![&*self.child]
-    }
-
-    fn children_mut(&mut self) -> &mut [Box<dyn Widget>] {
-        std::slice::from_mut(&mut self.child)
-    }
-
     fn traverse(&self, f: &mut dyn FnMut(&dyn Widget)) {
         f(self.child.as_ref());
         self.child.traverse(f);
@@ -116,15 +108,5 @@ mod test {
 
         button.traverse(&mut |widget: &dyn Widget| assert_eq!(id, widget.id()));
         button.traverse_mut(&mut |widget: &mut dyn Widget| assert_eq!(id, widget.id()));
-    }
-
-    #[test]
-    fn expose_children() {
-        FONT.set(init_font()).unwrap();
-        let text = Text::new("Hello");
-        let id = text.id();
-
-        let button = Button::new(text);
-        assert_eq!(button.children()[0].id(), id);
     }
 }